Fixing issues due to using newer tensorflow

+++ infogan/models/regularized_gan.py
@@ -57,8 +57,8 @@ class RegularizedGAN(object):
                      custom_fully_connected(image_size / 4 * image_size / 4 * 128).
                      fc_batch_norm().
                      apply(tf.nn.relu).
-                     reshape([-1, image_size / 4, image_size / 4, 128]).
-                     custom_deconv2d([0, image_size / 2, image_size / 2, 64], k_h=4, k_w=4).
+                     reshape([-1, int(image_size / 4), int(image_size / 4), 128]).
+                     custom_deconv2d([0, int(image_size / 2), int(image_size / 2), 64], k_h=4, k_w=4).
                      conv_batch_norm().
                      apply(tf.nn.relu).
                      custom_deconv2d([0] + list(image_shape), k_h=4, k_w=4).
